feat: make ui factory
