+++ src/CiscoRoomOsCodec.cs
@@ -1,9 +1,6 @@
-﻿using System;using System.Collections.Generic;using System.Globalization;using System.Linq;using System.Text;using System.Text.RegularExpressions;using Crestron.SimplSharp;using Crestron.SimplSharp.CrestronIO;using Crestron.SimplSharpPro.DeviceSupport;using Crestron.SimplSharpPro.Diagnostics;using Newtonsoft.Json;using Newtonsoft.Json.Linq;using PepperDash.Core;using PepperDash.Core.Intersystem.Tokens;using PepperDash.Core.Intersystem;using PepperDash.Essentials.Core;using PepperDash.Essentials.Core.Bridges;using PepperDash.Essentials.Core.Config;using PepperDash.Essentials.Core.DeviceInfo;using PepperDash.Essentials.Core.DeviceTypeInterfaces;using PepperDash.Essentials.Core.Routing;using PepperDash.Essentials.Devices.Common.Cameras;using PepperDash.Essentials.Devices.Common.Codec;using PepperDash.Essentials.Devices.Common.VideoCodec;using PepperDash.Essentials.Core.Queues;using Feedback = PepperDash.Essentials.Core.Feedback;using Serilog.Events;using Independentsoft.Exchange;using epi_videoCodec_ciscoExtended.UserInterfaceExtensions;using PepperDash.Essentials.Devices.Common.VideoCodec.Interfaces;using static Crestron.SimplSharpPro.DM.Audio;namespace epi_videoCodec_ciscoExtended{
+﻿using System;using System.Collections.Generic;using System.Globalization;using System.Linq;using System.Text;using System.Text.RegularExpressions;using Crestron.SimplSharp;using Crestron.SimplSharp.CrestronIO;using Crestron.SimplSharpPro.DeviceSupport;using Crestron.SimplSharpPro.Diagnostics;using Newtonsoft.Json;using Newtonsoft.Json.Linq;using PepperDash.Core;using PepperDash.Core.Intersystem.Tokens;using PepperDash.Core.Intersystem;using PepperDash.Essentials.Core;using PepperDash.Essentials.Core.Bridges;using PepperDash.Essentials.Core.Config;using PepperDash.Essentials.Core.DeviceInfo;using PepperDash.Essentials.Core.DeviceTypeInterfaces;using PepperDash.Essentials.Core.Routing;using PepperDash.Essentials.Devices.Common.Cameras;using PepperDash.Essentials.Devices.Common.Codec;using PepperDash.Essentials.Devices.Common.VideoCodec;using PepperDash.Essentials.Core.Queues;using Feedback = PepperDash.Essentials.Core.Feedback;using Serilog.Events;using PepperDash.Essentials.Devices.Common.VideoCodec.Interfaces;using static Crestron.SimplSharpPro.DM.Audio;using epi_videoCodec_ciscoExtended.UserInterface.UserInterfaceExtensions;namespace epi_videoCodec_ciscoExtended{
 
-
-
-
-	#region enums                                        internal enum eCommandType	{		SessionStart,		SessionEnd,		Command,		GetStatus,		GetConfiguration	};	public enum eExternalSourceType	{		camera,		desktop,		document_camera,		mediaplayer,		PC,		whiteboard,		other	}	public enum eExternalSourceMode	{		Ready,		NotReady,		Hidden,		Error	}	public enum eCodecPresentationStates	{		LocalOnly,		LocalRemote	}	public enum eCameraTrackingCapabilities	{		None,		SpeakerTrack,		PresenterTrack,		Both	}
+    #region enums                                                                            internal enum eCommandType	{		SessionStart,		SessionEnd,		Command,		GetStatus,		GetConfiguration	};	public enum eExternalSourceType	{		camera,		desktop,		document_camera,		mediaplayer,		PC,		whiteboard,		other	}	public enum eExternalSourceMode	{		Ready,		NotReady,		Hidden,		Error	}	public enum eCodecPresentationStates	{		LocalOnly,		LocalRemote	}	public enum eCameraTrackingCapabilities	{		None,		SpeakerTrack,		PresenterTrack,		Both	}
 
 
 
@@ -12,7 +9,7 @@
 IHasScheduleAwareness, IOccupancyStatusProvider, IHasCodecLayoutsAvailable, IHasCodecSelfView,
 ICommunicationMonitor, IRouting, IHasCodecCameras, IHasCameraAutoMode, IHasCodecRoomPresets,
 IHasExternalSourceSwitching, IHasBranding, IHasCameraOff, IHasCameraMute, IHasDoNotDisturbMode,
-IHasHalfWakeMode, IHasCallHold, IJoinCalls, IDeviceInfoProvider, IHasPhoneDialing, IVideoCodecUiExtensions	{		public event EventHandler<AvailableLayoutsChangedEventArgs> AvailableLayoutsChanged;		public event EventHandler<CurrentLayoutChangedEventArgs> CurrentLayoutChanged;		private event EventHandler<MinuteChangedEventArgs> MinuteChanged;		public event EventHandler<CodecInfoChangedEventArgs> CodecInfoChanged;		public event EventHandler<CameraTrackingCapabilitiesArgs> CameraTrackingCapabilitiesChanged;		public eCameraTrackingCapabilities CameraTrackingCapabilities { get; private set; }		private CTimer _scheduleCheckTimer;		private DateTime _scheduleCheckLast;		public Meeting ActiveMeeting { get; private set; }		private const int XSigEncoding = 28591;		public bool EndAllCallsOnMeetingJoin;		private List<Meeting> _currentMeetings;		private StringBuilder _feedbackListMessage;		private bool _feedbackListMessageIncoming;		private int _selectedPreset;		private bool _IsInPresentation;		private MediaChannelStatus _incomingPresentation;		[Flags]		public enum MediaChannelStatus		{			Unknown = 0,			None = 1,			Outgoing = 2,			Incoming = 4,			Video = 8,			Audio = 16,			Main = 32,			Presentation = 64,			OutgoingPresentation = 66,			IncomingPresentation = 68		}		private readonly Version _testedCodecFirmware = new Version("10.11.5.2");		private readonly Version _enhancedLayoutsFirmware = new Version("9.15.10.8");		private readonly Version _regressionFirmware = new Version("9.15.3.26");		private readonly Version _zoomDialFeatureFirmware = new Version("11.1.0.0");		public Version CodecFirmware { get; private set; }		private bool EnhancedLayouts		{			get			{				if (CodecFirmware == null) return false;				var returnValue = CodecFirmware.CompareTo(_enhancedLayoutsFirmware) >= 0;				Debug.Console(1, this, "Enhanced Layout Functionality is {0}.", returnValue ? "enabled" : "disabled");				return (returnValue);			}		}		private bool IsRegressionFirmware		{			get			{				if (CodecFirmware == null) return false;				var returnValue = CodecFirmware.CompareTo(_regressionFirmware) >= 0;				Debug.Console(1, this, "Currently running {0} firmware.", returnValue ? "current" : "legacy");				return (returnValue);			}		}		private bool ZoomDialerFirmware		{			get			{				var returnValue = FirmwareCompare(_zoomDialFeatureFirmware);				Debug.Console(1, this, "Enhanced Zoom Dialer Functionality is {0}.", returnValue ? "enabled" : "disabled");				return (returnValue);			}		}		public readonly WebexPinRequestHandler WebexPinRequestHandler;		public readonly DoNotDisturbHandler DoNotDisturbHandler;		public readonly UIExtensionsHandler UIExtensionsHandler;		public readonly UserInterfaceExtensions.ICiscoCodecUiExtensions UiExtensions;		private Meeting _currentMeeting;		private bool _standbyIsOn;		private bool _presentationActive;		private readonly bool _phonebookAutoPopulate;		private bool _phonebookInitialSearch;		private string _lastSearched;		private CiscoCodecConfig _config;		private readonly int _joinableCooldownSeconds;		public string ZoomMeetingId { get; private set; }		public string ZoomMeetingPasscode { get; private set; }		public string ZoomMeetingCommand { get; private set; }		public string ZoomMeetingHostKey { get; private set; }		public string ZoomMeetingReservedCode { get; private set; }		public string ZoomMeetingDialCode { get; private set; }		public string ZoomMeetingIp { get; private set; }		public string WebexMeetingNumber { get; private set; }		public string WebexMeetingRole { get; private set; }		public string WebexMeetingPin { get; private set; }		public eCodecPresentationStates PresentationStates;		private bool _externalSourceChangeRequested;		public bool Room { get; private set; }		public event EventHandler<DirectoryEventArgs> DirectoryResultReturned;		private CTimer _brandingTimer;		private CTimer _registrationCheckTimer;		public CommunicationGather PortGather { get; private set; }		public StatusMonitorBase CommunicationMonitor { get; private set; }		private readonly GenericQueue _receiveQueue;		public BoolFeedback PresentationViewMaximizedFeedback { get; private set; }		public BoolFeedback PresentationViewMinimizedFeedback { get; private set; }		public BoolFeedback PresentationViewDefaultFeedback { get; private set; }		public FeedbackGroup PresentationViewFeedbackGroup { get; private set; }		private string _currentPresentationView;		private eCameraTrackingCapabilities PreferredTrackingMode { get; set; }		public BoolFeedback RoomIsOccupiedFeedback { get; private set; }		public IntFeedback PeopleCountFeedback { get; private set; }		public BoolFeedback SelfviewIsOnFeedback { get; private set; }		public StringFeedback SelfviewPipPositionFeedback { get; private set; }		public StringFeedback LocalLayoutFeedback { get; private set; }		public BoolFeedback PresentationActiveFeedback { get; private set; }		public bool SpeakerTrackAvailability { get; private set; }		public bool SpeakerTrackStatus { get; private set; }		public bool PresenterTrackAvailability { get; private set; }		public bool PresenterTrackStatus { get; private set; }		public string PresenterTrackStatusName { get; private set; }		private string _currentLayoutBacker;		private string CurrentLayout		{			get { return _currentLayoutBacker != "Grid" ? _currentLayoutBacker : "Side by Side"; }			set { _currentLayoutBacker = value; }		}		public StringFeedback AvailableLayoutsFeedback { get; private set; }		public BoolFeedback LocalLayoutIsProminentFeedback { get; private set; }		public BoolFeedback FarEndIsSharingContentFeedback { get; private set; }
+IHasHalfWakeMode, IHasCallHold, IJoinCalls, IDeviceInfoProvider, IHasPhoneDialing	{		public event EventHandler<AvailableLayoutsChangedEventArgs> AvailableLayoutsChanged;		public event EventHandler<CurrentLayoutChangedEventArgs> CurrentLayoutChanged;		private event EventHandler<MinuteChangedEventArgs> MinuteChanged;		public event EventHandler<CodecInfoChangedEventArgs> CodecInfoChanged;		public event EventHandler<CameraTrackingCapabilitiesArgs> CameraTrackingCapabilitiesChanged;		public eCameraTrackingCapabilities CameraTrackingCapabilities { get; private set; }		private CTimer _scheduleCheckTimer;		private DateTime _scheduleCheckLast;		public Meeting ActiveMeeting { get; private set; }		private const int XSigEncoding = 28591;		public bool EndAllCallsOnMeetingJoin;		private List<Meeting> _currentMeetings;		private StringBuilder _feedbackListMessage;		private bool _feedbackListMessageIncoming;		private int _selectedPreset;		private bool _IsInPresentation;		private MediaChannelStatus _incomingPresentation;		[Flags]		public enum MediaChannelStatus		{			Unknown = 0,			None = 1,			Outgoing = 2,			Incoming = 4,			Video = 8,			Audio = 16,			Main = 32,			Presentation = 64,			OutgoingPresentation = 66,			IncomingPresentation = 68		}		private readonly Version _testedCodecFirmware = new Version("10.11.5.2");		private readonly Version _enhancedLayoutsFirmware = new Version("9.15.10.8");		private readonly Version _regressionFirmware = new Version("9.15.3.26");		private readonly Version _zoomDialFeatureFirmware = new Version("11.1.0.0");		public Version CodecFirmware { get; private set; }		private bool EnhancedLayouts		{			get			{				if (CodecFirmware == null) return false;				var returnValue = CodecFirmware.CompareTo(_enhancedLayoutsFirmware) >= 0;				Debug.Console(1, this, "Enhanced Layout Functionality is {0}.", returnValue ? "enabled" : "disabled");				return (returnValue);			}		}		private bool IsRegressionFirmware		{			get			{				if (CodecFirmware == null) return false;				var returnValue = CodecFirmware.CompareTo(_regressionFirmware) >= 0;				Debug.Console(1, this, "Currently running {0} firmware.", returnValue ? "current" : "legacy");				return (returnValue);			}		}		private bool ZoomDialerFirmware		{			get			{				var returnValue = FirmwareCompare(_zoomDialFeatureFirmware);				Debug.Console(1, this, "Enhanced Zoom Dialer Functionality is {0}.", returnValue ? "enabled" : "disabled");				return (returnValue);			}		}		public readonly WebexPinRequestHandler WebexPinRequestHandler;		public readonly DoNotDisturbHandler DoNotDisturbHandler;		public readonly UIExtensionsHandler UIExtensionsHandler;		private Meeting _currentMeeting;		private bool _standbyIsOn;		private bool _presentationActive;		private readonly bool _phonebookAutoPopulate;		private bool _phonebookInitialSearch;		private string _lastSearched;		private CiscoCodecConfig _config;		private readonly int _joinableCooldownSeconds;		public string ZoomMeetingId { get; private set; }		public string ZoomMeetingPasscode { get; private set; }		public string ZoomMeetingCommand { get; private set; }		public string ZoomMeetingHostKey { get; private set; }		public string ZoomMeetingReservedCode { get; private set; }		public string ZoomMeetingDialCode { get; private set; }		public string ZoomMeetingIp { get; private set; }		public string WebexMeetingNumber { get; private set; }		public string WebexMeetingRole { get; private set; }		public string WebexMeetingPin { get; private set; }		public eCodecPresentationStates PresentationStates;		private bool _externalSourceChangeRequested;		public bool Room { get; private set; }		public event EventHandler<DirectoryEventArgs> DirectoryResultReturned;		private CTimer _brandingTimer;		private CTimer _registrationCheckTimer;		public CommunicationGather PortGather { get; private set; }		public StatusMonitorBase CommunicationMonitor { get; private set; }		private readonly GenericQueue _receiveQueue;		public BoolFeedback PresentationViewMaximizedFeedback { get; private set; }		public BoolFeedback PresentationViewMinimizedFeedback { get; private set; }		public BoolFeedback PresentationViewDefaultFeedback { get; private set; }		public FeedbackGroup PresentationViewFeedbackGroup { get; private set; }		private string _currentPresentationView;		private eCameraTrackingCapabilities PreferredTrackingMode { get; set; }		public BoolFeedback RoomIsOccupiedFeedback { get; private set; }		public IntFeedback PeopleCountFeedback { get; private set; }		public BoolFeedback SelfviewIsOnFeedback { get; private set; }		public StringFeedback SelfviewPipPositionFeedback { get; private set; }		public StringFeedback LocalLayoutFeedback { get; private set; }		public BoolFeedback PresentationActiveFeedback { get; private set; }		public bool SpeakerTrackAvailability { get; private set; }		public bool SpeakerTrackStatus { get; private set; }		public bool PresenterTrackAvailability { get; private set; }		public bool PresenterTrackStatus { get; private set; }		public string PresenterTrackStatusName { get; private set; }		private string _currentLayoutBacker;		private string CurrentLayout		{			get { return _currentLayoutBacker != "Grid" ? _currentLayoutBacker : "Side by Side"; }			set { _currentLayoutBacker = value; }		}		public StringFeedback AvailableLayoutsFeedback { get; private set; }		public BoolFeedback LocalLayoutIsProminentFeedback { get; private set; }		public BoolFeedback FarEndIsSharingContentFeedback { get; private set; }
 
 
 
@@ -53,14 +50,11 @@ IHasHalfWakeMode, IHasCallHold, IJoinCalls, IDeviceInfoProvider, IHasPhoneDialin
 
 		//public Action<UiWebViewDisplayActionArgs> UiWebViewDisplayAction { get; set; }
 		//public event EventHandler<UiExtensionsPanelEventArgs> UiExtensionsPanelClickEvent;
-		public IVideoCodecUiExtensionsHandler VideoCodecUiExtensionsHandler { get; set; }
 		private readonly IBasicCommunication _comms;
 		// Constructor for IBasicCommunication
 		public CiscoCodec(DeviceConfig config, IBasicCommunication comm)
 : base(config)		{			CodecStatus = new CiscoCodecStatus.RootObject();			DeviceInfo = new DeviceInfo();			CodecInfo = new CiscoCodecInfo(this);			_lastSearched = string.Empty;			_phonebookInitialSearch = true;			CurrentLayout = string.Empty;			_receiveQueue = new GenericQueue(Key + "-queue", 500);			WebexPinRequestHandler = new WebexPinRequestHandler(this, comm, _receiveQueue);			DoNotDisturbHandler = new DoNotDisturbHandler(this, comm, _receiveQueue);			UIExtensionsHandler = new UIExtensionsHandler(this, comm, _receiveQueue);			_comms = comm;			CrestronEnvironment.ProgramStatusEventHandler += a =>			{				if (a != eProgramStatusEventType.Stopping) return;				EndGracefully();			};			CrestronEnvironment.SystemEventHandler += a =>			{				if (a != eSystemEventType.Rebooting) return;				EndGracefully();			};			var props = JsonConvert.DeserializeObject<CiscoCodecConfig>(config.Properties.ToString());
-
-			UiExtensions = props.Extensions;
-			VideoCodecUiExtensionsHandler = new UiExtensionsHandler(this, _comms, EnqueueCommand);			_scheduleCheckTimer = new CTimer(ScheduleTimeCheck, null, 0, 15000);			_config = props;			MeetingsToDisplay = _config.OverrideMeetingsLimit ? 50 : 0;			_timeFormatSpecifier = _config.TimeFormatSpecifier ?? "t";			_dateFormatSpecifier = _config.DateFormatSpecifier ?? "d";			_joinableCooldownSeconds = _config.JoinableCooldownSeconds;			EndAllCallsOnMeetingJoin = _config.EndAllCallsOnMeetingJoin;			if (_config.Sharing != null)			{				PresentationStates = _config.Sharing.DefaultShareLocalOnly					? eCodecPresentationStates.LocalOnly					: eCodecPresentationStates.LocalRemote;			}			else			{				PresentationStates = eCodecPresentationStates.LocalRemote;			}			PreferredTrackingMode = eCameraTrackingCapabilities.SpeakerTrack;			var trackingMode = _config.DefaultCameraTrackingMode ?? string.Empty;			if (!String.IsNullOrEmpty(trackingMode))			{				if (trackingMode.Contains("presenter"))				{					PreferredTrackingMode = eCameraTrackingCapabilities.PresenterTrack;				}			}
+			_scheduleCheckTimer = new CTimer(ScheduleTimeCheck, null, 0, 15000);			_config = props;			MeetingsToDisplay = _config.OverrideMeetingsLimit ? 50 : 0;			_timeFormatSpecifier = _config.TimeFormatSpecifier ?? "t";			_dateFormatSpecifier = _config.DateFormatSpecifier ?? "d";			_joinableCooldownSeconds = _config.JoinableCooldownSeconds;			EndAllCallsOnMeetingJoin = _config.EndAllCallsOnMeetingJoin;			if (_config.Sharing != null)			{				PresentationStates = _config.Sharing.DefaultShareLocalOnly					? eCodecPresentationStates.LocalOnly					: eCodecPresentationStates.LocalRemote;			}			else			{				PresentationStates = eCodecPresentationStates.LocalRemote;			}			PreferredTrackingMode = eCameraTrackingCapabilities.SpeakerTrack;			var trackingMode = _config.DefaultCameraTrackingMode ?? string.Empty;			if (!String.IsNullOrEmpty(trackingMode))			{				if (trackingMode.Contains("presenter"))				{					PreferredTrackingMode = eCameraTrackingCapabilities.PresenterTrack;				}			}
 
 			// Use the configured phonebook results limit if present
 			if (props.PhonebookResultsLimit > 0)			{				_phonebookResultsLimit = props.PhonebookResultsLimit;			}
