--- src/UserInterfaceExtensions/ICiscoCodecUiExtensionsHandler.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace epi_videoCodec_ciscoExtended.UserInterfaceExtensions
-{
-	public interface ICiscoCodecUiExtensionsHandler
-	{
-		/// <summary>
-		/// Called by receive parser to parse event feedback
-		/// </summary>
-		/// <param name="panel"></param>
-		void ParseStatus(Panels.CiscoCodecEvents.Panel panel);
-	}
-}
